perf(op-log): lazy-load native SQLite backend off the web/desktop bundle
The DI token statically imported native-sqlite-backend, pulling the whole SQLite
graph (the @capacitor-community/sqlite plugin, the SQLite adapter, value codec and
migration) into the eager bundle on EVERY platform — including web/PWA/Electron,
which never use it. The runtime gate can't be tree-shaken, so the code shipped to
all users and tripped Lighthouse's script-count budget (202 > 200).

Move the cheap platform gate into a tiny native-sqlite-gate module (only
@capacitor/core, already bundled) so the token decides synchronously, and
dynamic-import the heavy backend only when the gate passes (Android). The backend
is still resolved once and shared across both stores; each store gets its adapter
synchronously via NativeOpLogAdapter, whose async init awaits the shared resolver.

No behaviour change: bootstrap/migration/fallback logic is relocated, not altered.
601 op-log persistence specs green.
