+++ src/app/op-log/persistence/op-log-db-adapter.token.ts
@@ -1,10 +1,8 @@
 import { InjectionToken } from '@angular/core';
 import { OpLogDbAdapter } from './op-log-db-adapter';
 import { IndexedDbOpLogAdapter } from './indexed-db-op-log-adapter';
-import {
-  createNativeSqliteOpLogAdapterFactory,
-  shouldUseNativeSqliteOpLogBackend,
-} from './native-sqlite-backend';
+import { NativeOpLogAdapter } from './native-op-log-adapter';
+import { shouldUseNativeSqliteOpLogBackend } from './native-sqlite-gate';
 
 /**
  * Factory that produces a fresh {@link OpLogDbAdapter}.
@@ -16,6 +14,26 @@ import {
  * factory, not an instance.
  */
 export type OpLogDbAdapterFactory = () => OpLogDbAdapter;
+
+/**
+ * Build the lazy native-SQLite factory. The heavy SQLite graph (the
+ * `@capacitor-community/sqlite` plugin, the SQLite adapter, the value codec, the
+ * migration) lives behind a dynamic `import()` so web/PWA/Electron — which never
+ * pass the gate — do not bundle it. The backend is resolved once and shared
+ * across both stores (one connection, one bootstrap). Each store still gets its
+ * adapter synchronously: a {@link NativeOpLogAdapter} whose async `init()` awaits
+ * the shared, memoized resolver.
+ */
+const createLazyNativeSqliteFactory = (): OpLogDbAdapterFactory => {
+  let resolver: Promise<() => Promise<OpLogDbAdapter>> | undefined;
+  const resolveBackend = (): Promise<OpLogDbAdapter> => {
+    resolver ??= import('./native-sqlite-backend').then((m) =>
+      m.createSharedNativeBackendResolver(),
+    );
+    return resolver.then((resolve) => resolve());
+  };
+  return () => new NativeOpLogAdapter(resolveBackend);
+};
 
 /**
  * DI seam for the op-log persistence backend (Phase B of the SQLite migration;
@@ -34,7 +52,7 @@ export const OP_LOG_DB_ADAPTER_FACTORY = new InjectionToken<OpLogDbAdapterFactor
     providedIn: 'root',
     factory: () =>
       shouldUseNativeSqliteOpLogBackend()
-        ? createNativeSqliteOpLogAdapterFactory()
+        ? createLazyNativeSqliteFactory()
         : () => new IndexedDbOpLogAdapter(),
   },
 );
